refactor(state): move data's state component to separate file
