+++ src/pages/data/index.js
@@ -1,59 +1,11 @@
 import React from 'react'
-import { graphql, Link } from 'gatsby'
-import marked from 'marked'
+import { graphql } from 'gatsby'
 import { Flex, Box } from '../../components/common/flexbox'
-import { UnstyledList } from '../../components/common/lists'
+import State from '../../components/common/state-data'
 import Layout from '../../components/layout'
 import { SyncInfobox } from '../../components/common/infobox'
-import slug from '../../utilities/slug'
 import SummaryTable from '../../components/common/summary-table'
 import '../../scss/pages/data.scss'
-
-const State = ({ state }) => (
-  <>
-    <h3>
-      <Link to={`/data/state/${slug(state.name)}`}>{state.name}</Link>
-    </h3>
-    <SummaryTable data={state.stateData} />
-    <UnstyledList>
-      {state.twitter && (
-        <li>
-          <a href={`https://twitter.com/${state.twitter}`}>
-            {state.name} on Twitter
-          </a>
-        </li>
-      )}
-      {state.covid19Site && (
-        <li>
-          <a href={state.covid19Site}>
-            Best current data source for {state.name}
-          </a>
-        </li>
-      )}
-      <li>
-        <Link to={`/data/state/${slug(state.name)}#historical`}>
-          Historical data for {state.name}
-        </Link>
-      </li>
-      {state.stateData.grade && (
-        <li>
-          Data quality grade:{' '}
-          <span className="state-grade">{state.stateData.grade}</span>
-        </li>
-      )}
-    </UnstyledList>
-    {state.notes && (
-      <div
-        dangerouslySetInnerHTML={{
-          __html: marked(state.notes),
-        }}
-      />
-    )}
-    <a className="top-link" href="#states-top" title="top">
-      ↑ (return to top)
-    </a>
-  </>
-)
 
 const StateList = ({ states, stateData }) => {
   const stateList = []
